+++ packages/banner/src/Banner.tsx
@@ -5,8 +5,9 @@ import WarningIcon from '@leafygreen-ui/icon/dist/Warning';
 import CheckmarkWithCircleIcon from '@leafygreen-ui/icon/dist/CheckmarkWithCircle';
 import XIcon from '@leafygreen-ui/icon/dist/X';
 import { css, cx } from '@leafygreen-ui/emotion';
-import { uiColors } from '@leafygreen-ui/palette';
+import { palette } from '@leafygreen-ui/palette';
 import { HTMLElementProps } from '@leafygreen-ui/lib';
+import IconButton from '@leafygreen-ui/icon-button';
 
 const Variant = {
   Info: 'info',
@@ -25,33 +26,31 @@ const baseBannerStyles = css`
   position: relative;
   display: flex;
   min-height: 40px;
-  padding: 9px 12px 9px 20px;
+  padding: 10px 12px 10px 20px;
   border-width: 1px 1px 1px 0px;
   border-style: solid;
-  border-radius: 6px;
-  font-size: 14px;
+  border-radius: 12px;
+  font-size: 13px;
   line-height: 20px;
+  font-family: 'Euclid Circular A', Helvetica, Arial, sans-serif; // TODO: Refresh – remove when fonts are updated
 
   &:before {
     content: '';
     position: absolute;
-    width: 6px;
+    width: 13px;
     top: -1px;
     bottom: -1px;
     left: 0px;
-    border-radius: 6px 0px 0px 6px;
+    border-radius: 12px 0px 0px 12px;
   }
+`;
 
-  &:after {
-    content: '';
-    position: absolute;
-    left: 4px;
-    top: -1px;
-    bottom: -1px;
-    width: 2px;
-    border-top: 1px solid;
-    border-bottom: 1px solid;
-    border-radius: 0.5px 0px 0px 0.5px;
+//TODO: refresh - remove when IconButton focus styles are updated
+const baseDismissButtonStyles = css`
+  &:focus {
+    color: ${palette.gray.dark3};
+    outline: 2px solid ${palette.blue.light1};
+    border: 2px solid ${palette.white};
   }
 `;
 
@@ -73,93 +72,160 @@ const renderedImageStyles = css`
   flex-shrink: 0;
 `;
 
-const bannerVariantStyles: Record<Variant, string> = {
-  [Variant.Info]: css`
-    color: ${uiColors.blue.dark2};
-    border-color: ${uiColors.blue.light2};
-    border-left-color: ${uiColors.blue.base};
-    background-color: ${uiColors.blue.light3};
+type StyledElements = 'body' | 'dismissButton';
 
-    &:before {
-      background-color: ${uiColors.blue.base};
-    }
+const bannerVariantStyles: Record<Variant, Record<StyledElements, string>> = {
+  [Variant.Info]: {
+    body: css`
+      color: ${palette.blue.dark2};
+      border-color: ${palette.blue.light2};
+      border-left-color: ${palette.blue.base};
+      background-color: ${palette.blue.light3};
 
-    &:after {
-      border-color: ${uiColors.blue.light2};
-      background-color: ${uiColors.blue.light3};
-    }
-  `,
+      &:before {
+        background: linear-gradient(
+          to left,
+          transparent 6px,
+          ${palette.blue.base} 6px
+        );
+      }
+    `,
+    dismissButton: css`
+      color: ${palette.blue.dark2};
 
-  [Variant.Warning]: css`
-    color: ${uiColors.yellow.dark2};
-    border-color: ${uiColors.yellow.light2};
-    border-left-color: ${uiColors.yellow.base};
-    background-color: ${uiColors.yellow.light3};
+      &:active,
+      &:hover {
+        color: ${palette.blue.dark2};
 
-    &:before {
-      background-color: ${uiColors.yellow.base};
-    }
+        &:before {
+          background-color: ${palette.blue.light2};
+        }
+      }
 
-    &:after {
-      border-color: ${uiColors.yellow.light2};
-      background-color: ${uiColors.yellow.light3};
-    }
-  `,
+      &:focus {
+        &:before {
+          background-color: ${palette.blue.light2};
+        }
+      }
+    `,
+  },
 
-  [Variant.Danger]: css`
-    color: ${uiColors.red.dark2};
-    border-color: ${uiColors.red.light2};
-    border-left-color: ${uiColors.red.base};
-    background-color: ${uiColors.red.light3};
+  [Variant.Warning]: {
+    body: css`
+      color: ${palette.yellow.dark2};
+      border-color: ${palette.yellow.light2};
+      border-left-color: ${palette.yellow.base};
+      background-color: ${palette.yellow.light3};
 
-    &:before {
-      background-color: ${uiColors.red.base};
-    }
+      &:before {
+        background: linear-gradient(
+          to left,
+          transparent 6px,
+          ${palette.yellow.base} 6px
+        );
+      }
+    `,
+    dismissButton: css`
+      color: ${palette.yellow.dark2};
 
-    &:after {
-      border-color: ${uiColors.red.light2};
-      background-color: ${uiColors.red.light3};
-    }
-  `,
+      &:active,
+      &:hover {
+        color: ${palette.yellow.dark2};
 
-  [Variant.Success]: css`
-    color: ${uiColors.green.dark2};
-    border-color: ${uiColors.green.light2};
-    border-left-color: ${uiColors.green.base};
-    background-color: ${uiColors.green.light3};
+        &:before {
+          background-color: ${palette.yellow.light2};
+        }
+      }
 
-    &:before {
-      background-color: ${uiColors.green.base};
-    }
+      &:focus {
+        &:before {
+          background-color: ${palette.yellow.light2};
+        }
+      }
+    `,
+  },
 
-    &:after {
-      border-color: ${uiColors.green.light2};
-      background-color: ${uiColors.green.light3};
-    }
-  `,
+  [Variant.Danger]: {
+    body: css`
+      color: ${palette.red.dark2};
+      border-color: ${palette.red.light2};
+      border-left-color: ${palette.red.base};
+      background-color: ${palette.red.light3};
+
+      &:before {
+        background: linear-gradient(
+          to left,
+          transparent 6px,
+          ${palette.red.base} 6px
+        );
+      }
+    `,
+    dismissButton: css`
+      color: ${palette.red.dark2};
+
+      &:active,
+      &:hover {
+        color: ${palette.red.dark2};
+
+        &:before {
+          background-color: ${palette.red.light2};
+        }
+      }
+
+      &:focus {
+        &:before {
+          background-color: ${palette.red.light2};
+        }
+      }
+    `,
+  },
+
+  [Variant.Success]: {
+    body: css`
+      color: ${palette.green.dark2};
+      border-color: ${palette.green.light2};
+      border-left-color: ${palette.green.base};
+      background-color: ${palette.green.light3};
+
+      &:before {
+        background: linear-gradient(
+          to left,
+          transparent 6px,
+          ${palette.green.dark1} 6px
+        );
+      }
+    `,
+    dismissButton: css`
+      color: ${palette.green.dark2};
+
+      &:active,
+      &:hover {
+        color: ${palette.green.dark2};
+
+        &:before {
+          background-color: ${palette.green.light2};
+        }
+      }
+
+      &:focus {
+        &:before {
+          background-color: ${palette.green.light2};
+        }
+      }
+    `,
+  },
 } as const;
 
 const map = {
-  [Variant.Info]: { color: uiColors.blue.base, icon: InfoWithCircleIcon },
+  [Variant.Info]: { color: palette.blue.base, icon: InfoWithCircleIcon },
   [Variant.Warning]: {
-    color: uiColors.yellow.dark2,
+    color: palette.yellow.dark2,
     icon: ImportantWithCircleIcon,
   },
-  [Variant.Danger]: { color: uiColors.red.base, icon: WarningIcon },
+  [Variant.Danger]: { color: palette.red.base, icon: WarningIcon },
   [Variant.Success]: {
-    color: uiColors.green.dark1,
+    color: palette.green.dark1,
     icon: CheckmarkWithCircleIcon,
-  },
-};
-
-const dismissibleMap = {
-  [Variant.Info]: { color: uiColors.blue.dark2 },
-  [Variant.Warning]: {
-    color: uiColors.yellow.dark2,
-  },
-  [Variant.Danger]: { color: uiColors.red.dark1 },
-  [Variant.Success]: {
-    color: uiColors.green.dark2,
   },
 };
 
@@ -175,13 +241,13 @@ const getTextStyle = (image: boolean, dismissible: boolean) => {
   };
 
   if (image) {
-    styleObj.marginLeft = '15px';
+    styleObj.marginLeft = '17px';
     styleObj.marginRight = '4px';
     if (dismissible) {
       styleObj.marginRight = `${defaultIconSize + defaultBorderSpacing}px`;
     }
   } else {
-    styleObj.marginLeft = `${15}px`;
+    styleObj.marginLeft = `13px`;
     styleObj.marginRight = '10px';
     if (dismissible) {
       styleObj.marginRight = `${defaultIconSize + 16}px`;
@@ -258,7 +324,7 @@ export default function Banner({
       className={cx(
         flexShrink,
         css`
-          margin-top: 1px;
+          margin-top: 3px;
         `,
       )}
     />
@@ -267,17 +333,32 @@ export default function Banner({
   return (
     <div
       role="alert"
-      className={cx(baseBannerStyles, bannerVariantStyles[variant], className)}
+      className={cx(
+        baseBannerStyles,
+        bannerVariantStyles[variant].body,
+        className,
+      )}
       {...rest}
     >
       {renderIcon}
       <div className={getTextStyle(image != null, dismissible)}>{children}</div>
       {dismissible && (
-        <XIcon
-          fill={dismissibleMap[variant].color}
+        <IconButton
+          className={cx(
+            flexShrink,
+            cursorPointer,
+            css`
+              margin-top: -3px;
+              left: 8px;
+            `,
+            bannerVariantStyles[variant].dismissButton,
+            baseDismissButtonStyles,
+          )}
+          aria-label="Close Message"
           onClick={onClose}
-          className={cx(flexShrink, cursorPointer)}
-        />
+        >
+          <XIcon />
+        </IconButton>
       )}
     </div>
   );