+++ packages/select/src/Option.tsx
@@ -6,8 +6,9 @@ import { usePrevious } from '@leafygreen-ui/hooks';
 import { createDataProp } from '@leafygreen-ui/lib';
 import CheckmarkIcon from '@leafygreen-ui/icon/dist/Checkmark';
 import { LGGlyph } from '@leafygreen-ui/icon/src/types';
-import { colorSets } from './styleSets';
+import { colorSets, Mode } from './styleSets';
 import SelectContext from './SelectContext';
+import { fontFamilies } from '@leafygreen-ui/tokens';
 
 const option = createDataProp('option');
 
@@ -20,11 +21,11 @@ export type ReactEmpty = null | undefined | false | '';
 const optionStyle = css`
   display: flex;
   width: 100%;
-  padding: 10px 12px;
   outline: none;
   overflow-wrap: anywhere;
   transition: background-color 150ms ease-in-out;
   position: relative;
+  padding: 8px 12px;
 
   &:before {
     content: '';
@@ -44,6 +45,7 @@ const optionStyle = css`
 const optionTextStyle = css`
   display: flex;
   align-items: center;
+  font-family: ${fontFamilies.default};
 `;
 
 const iconStyle = css`
@@ -91,6 +93,9 @@ export function InternalOption({
 
   const ref = useRef<HTMLLIElement>(null);
 
+  const showDeselectionStyle =
+    selected && (mode === Mode.Light || !isDeselection);
+
   const scrollIntoView = useCallback(() => {
     if (ref.current == null) {
       return null;
@@ -133,7 +138,11 @@ export function InternalOption({
       className={cx(optionTextStyle, {
         [css`
           font-weight: bold;
-        `]: selected && !isDeselection,
+        `]: showDeselectionStyle,
+        // TODO: Refresh - remove darkMode logic
+        [css`
+          font-family: ${fontFamilies.legacy};
+        `]: mode === Mode.Dark,
       })}
     >
       {children}
@@ -178,26 +187,25 @@ export function InternalOption({
     }
   }
 
-  const checkmark =
-    selected && !isDeselection ? (
-      <CheckmarkIcon
-        key="checkmark"
-        className={cx(
-          iconStyle,
-          css`
-            color: ${colorSet.icon.selected};
-          `,
-          {
-            [glyphFocusStyle]: showFocus,
-            [css`
-              color: ${colorSet.icon.disabled};
-            `]: disabled,
-          },
-        )}
-      />
-    ) : (
-      iconPlaceholder
-    );
+  const checkmark = showDeselectionStyle ? (
+    <CheckmarkIcon
+      key="checkmark"
+      className={cx(
+        iconStyle,
+        css`
+          color: ${colorSet.icon.selected};
+        `,
+        {
+          [glyphFocusStyle]: showFocus,
+          [css`
+            color: ${colorSet.icon.disabled};
+          `]: disabled,
+        },
+      )}
+    />
+  ) : (
+    iconPlaceholder
+  );
 
   let renderedChildren: React.ReactNode;
 
@@ -244,6 +252,10 @@ export function InternalOption({
           color: ${colorSet.text.base};
         `,
         {
+          // TODO: Refresh - remove dark mode conditional styles
+          [css`
+            padding: 10px 12px;
+          `]: mode === Mode.Dark,
           [css`
             &:hover {
               background-color: ${colorSet.background.hovered};