+++ packages/typography/src/styles.tsx
@@ -1,13 +1,20 @@
 import { css } from '@leafygreen-ui/emotion';
 
+// TODO: Refresh - Update this when DarkMode is designed
 export const typeScale1 = css`
-  font-size: 14px;
+  font-size: 13px;
   line-height: 20px;
   letter-spacing: 0px;
 `;
 
 export const typeScale2 = css`
   font-size: 16px;
+  line-height: 24px;
+  letter-spacing: 0px;
+`;
+
+export const codeTypeScale2 = css`
+  font-size: 15px;
   line-height: 24px;
   letter-spacing: 0px;
 `;