+++ packages/interaction-ring/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/interaction-ring
 
+## 2.0.0
+
+### Major Changes
+
+- 8457f92: Updates interaction ring styles in line with visual brand refresh
+
+### Patch Changes
+
+- Updated dependencies [cb54eef]
+  - @leafygreen-ui/palette@3.3.1
+
 ## 1.1.1
 
 ### Patch Changes